requirements.txt
langgraph
langchain-openai
tavily-python
arxiv
fal-client
python-dotenv